requirements.txt
requests
tqdm
lxml
beautifulsoup4
pandas
python-dotenv
langchain-text-splitters
openai
langchain-openai
chromadb
langchain-chroma
langchain
langchain-community
